+++ example.py
@@ -1,13 +1,16 @@
-from frewpy import FrewModel
 from pprint import pprint
+import time
+
+from frewpy import FrewModel
+
+
+start_time = time.time()
 
 file_path = r"C:\Users\Fred.White\Documents\Work\Frewpy\models\SLS B4 South Basement.fwd"
 model = FrewModel(file_path)
 model.analyse()
-results = model.wall.get_results()
-
-pprint(results)
-pprint(results[7])
-
-
+model.wall.plot_results()
 model.close()
+
+delta_time = time.time() - start_time
+print(f'Executed in {delta_time:.2f} sec')